pico: update for new toolchains and pico2 support
The micro-ROS common meta defines "-DRCUTILS_NO_64_ATOMIC=ON", which
enables atomic ops functions in micro-ros/rcutils. But pico-sdk has
build-in atomic ops, so they conflict and result in link errors.

So user meta is added to avoid the errors.

Latest pico-quick-toolchain 4.1.0 is used to have proper clock_gettime
support.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>